% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipcancer.R
\docType{data}
\name{smoking}
\alias{smoking}
\title{Smoking Cessation Data}
\description{
data set consists of 24 studies conducted to study smoking cessation interventions.
There are a total of four treatments considered in the 24 studies.  The treatments
are 1) No contact, 2) self help, 3) individual counseling, and 4) group counseling.
This results in six possible treatment comparisons
}
\format{
data: A data frame with 50 rows and the following variables:
\describe{
  \item{sid}{study identification number}
  \item{tid}{treatment identification number}
  \item{r}{number of subjects that successfully abstained from smoking}
  \item{n}{number of subjects in the study}
}
}
\source{
Hasselblad V. (1998) <doi:10.1177/0272989X9801800110>
}
